pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudrift-runners"
version = "0.1.0"
description = "Serverless GCP controller for ephemeral GitHub Actions runners on CloudRift VMs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "functions-framework>=3,<4",
    "requests>=2.28,<3",
    "google-cloud-firestore>=2,<3",
    "google-cloud-secret-manager>=2,<3",
    "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7,<9",
    "responses>=0.23,<1",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
